+++ src/ConfigProvider.php
@@ -6,7 +6,7 @@
 
 declare(strict_types=1);
 
-namespace Mezzio\Navigation;
+namespace Ikoss\Mezzio\Navigation;
 
 use Laminas\Navigation\Navigation;