+++ test/Service/MezzioNavigationAbstractServiceFactoryTest.php
@@ -6,11 +6,11 @@
 
 declare(strict_types=1);
 
-namespace MezzioTest\Navigation\Service;
+namespace IkossTest\Mezzio\Navigation\Service;
 
+use Ikoss\Mezzio\Navigation\Service\MezzioNavigationAbstractServiceFactory;
 use Laminas\Navigation\Navigation;
 use Mezzio\Helper\UrlHelper;
-use Mezzio\Navigation\Service\MezzioNavigationAbstractServiceFactory;
 use Mezzio\Router\LaminasRouter;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;